+++ src/c++/2d/heat_2d/packages/src/observation_stations.cpp
@@ -25,25 +25,25 @@
 #include "observation_stations.h"
 
 int  def_observation_stations(std::vector<std::string>& obs_stations, std::vector<_ObservationPoint>& obs_points, KDTree xy_tree, std::vector<double> x, std::vector<double> y,
-    double Lx, double Ly, double dx, double dy, int nx, int ny)
+    double Lx, double Ly, double dx, double dy, size_t nx, size_t ny)
 {
     int status;
     status = add_hard_coded_observation_points(obs_points, x, y, Lx, Ly, dx, dy, nx, ny);
 
     int nsig = 0;
-    for (int i = 0; i < obs_points.size(); ++i)
+    for (size_t i = 0; i < obs_points.size(); ++i)
     {
         nsig = std::max(nsig, std::max( (int)std::log10(obs_points[i].x), (int)std::log10(obs_points[i].x) ));
     }
     //nsig += 1;
     
-    for (int i = 0; i < obs_points.size(); ++i)
+    for (size_t i = 0; i < obs_points.size(); ++i)
     {
         obs_stations.push_back(setup_obs_name(obs_points[i].x, obs_points[i].y, nsig, obs_points[i].name));
     }
     
     // Find the index of the observation point
-    for (int i = 0; i < obs_points.size(); ++i)
+    for (size_t i = 0; i < obs_points.size(); ++i)
     {
         std::vector<double> point = {obs_points[i].x, obs_points[i].y};
         obs_points[i].idx = xy_tree.nearest_index(point);
@@ -51,11 +51,10 @@ int  def_observation_stations(std::vector<std::string>& obs_stations, std::vecto
     return status;
 }
 
-int add_hard_coded_observation_points(std::vector<_ObservationPoint>& obs_points, std::vector<double> x, std::vector<double> y, double Lx, double Ly, double dx, double dy, int nx, int ny)
+int add_hard_coded_observation_points(std::vector<_ObservationPoint>& obs_points, std::vector<double> x, std::vector<double> y, double Lx, double Ly, double dx, double dy, size_t nx, size_t ny)
 {
-
     // Initialize observation station locations
-    int ptr_obs = idx(nx / 2, ny / 2, ny);
+    size_t ptr_obs = idx(nx / 2, ny / 2, ny);
     obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Centre"));
 
     // boundaries --------------------------------------------------------------
@@ -84,56 +83,9 @@ int add_hard_coded_observation_points(std::vector<_ObservationPoint>& obs_points
     ptr_obs = idx(nx - 2, 1     , ny);
     obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "SE"));
 
-    // halfway stations --------------------------------------------------------
-    //ptr_obs  = idx(    nx / 4 + 1,     ny / 2    , ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Halfway W"));
-    //
-    //ptr_obs  = idx(3 * nx / 4 - 1,     ny / 2    , ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Halfway E"));
-    //
-    //ptr_obs  = idx(    nx / 2    ,     ny / 4 + 1, ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Halfway S"));
-    //
-    //ptr_obs  = idx(    nx / 2    , 3 * ny / 4 - 1, ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Halfway N"));
-    //
-    //ptr_obs = idx(    nx / 4 + 1,     ny / 4 + 1, ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Halfway SW"));
-    //
-    //ptr_obs = idx(3 * nx / 4 - 1, 3 * ny / 4 - 1, ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Halfway NE"));
-    //
-    //ptr_obs = idx(    nx / 4 + 1, 3 * ny / 4 - 1, ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Halfway NW"));
-    //
-    //ptr_obs = idx(3 * nx / 4 - 1,     ny / 4 + 1, ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "Halfway SE"));
-
-    // stations on equal distance of centre (ie 2500.0 [m]) ---------------------
-    //double x_a = std::min(Lx / 2., 2500.);
-    //double x_b = std::min(Lx / 2., 2000.);
-    //double x_c = std::min(Lx / 2., 1500.);
-    //double x_d = std::min(Lx / 2., 0.);
-    //double y_a = std::min(Ly / 2., 0.);
-    //double y_b = std::min(Ly / 2., 1500.);
-    //double y_c = std::min(Ly / 2., 2000.);
-    //double y_d = std::min(Ly / 2., 2500.);
-    //
-    //ptr_obs = idx(int((y_a / dy) + nx / 2), int((x_a / dx) + ny / 2), ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "D"));
-    //
-    //ptr_obs = idx(int((y_b / dy) + nx / 2), int((x_b / dx) + ny / 2), ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "C"));
-    //
-    //ptr_obs = idx(int((y_c / dy) + nx / 2), int((x_c / dx) + ny / 2), ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "B"));
-    //
-    //ptr_obs = idx(int((y_d / dy) + nx / 2), int((x_d / dx) + ny / 2), ny);
-    //obs_points.push_back(add_obs(ptr_obs, x[ptr_obs], y[ptr_obs], "A"));
-
     return 0;
 }
-struct _ObservationPoint add_obs(int ptr_obs, double x_obs, double y_obs, std::string obs_name)
+struct _ObservationPoint add_obs(size_t ptr_obs, double x_obs, double y_obs, std::string obs_name)
 {
     struct _ObservationPoint obs_point; 
     obs_point.name = obs_name;
@@ -166,7 +118,7 @@ std::string string_format_with_zeros(double value, int width)
     }
     return oss.str();
 }
-inline int idx(int i, int j, int ny)
+inline size_t idx(size_t i, size_t j, size_t ny)
 {
     return i * ny + j;
 }
